+++ sovtimer/views.py
@@ -10,6 +10,7 @@ from django.contrib.auth.decorators import login_required, permission_required
 
 from django.http import JsonResponse
 from django.shortcuts import render
+from django.utils.translation import gettext_lazy as _
 
 from sovtimer import __title__
 from sovtimer.app_settings import avoid_cdn
@@ -18,7 +19,7 @@ from allianceauth.services.hooks import get_extension_logger
 
 import datetime as dt
 
-from eveuniverse.models import EveRegion, EveConstellation, EveSolarSystem
+from eveuniverse.models import EveSolarSystem
 
 from sovtimer.providers import esi
 
@@ -68,8 +69,11 @@ def dashboard_data(request) -> JsonResponse:
             ).results()
 
             defender_name = alliance_esi["name"]
-            defender_name_html = '<a href="https://evemaps.dotlan.net/search?q={defender_name}" target="_blank" rel="noopener noreferer">{defender_name}</a>'.format(
-                defender_name=defender_name
+            defender_name_html = (
+                '<a href="https://evemaps.dotlan.net/search?q={defender_name}" '
+                'target="_blank" rel="noopener noreferer">{defender_name}</a>'.format(
+                    defender_name=defender_name
+                )
             )
 
             eve_solar_system = EveSolarSystem.objects.get(
@@ -77,18 +81,27 @@ def dashboard_data(request) -> JsonResponse:
             )
 
             solar_system_name = eve_solar_system.name
-            solar_system_name_html = '<a href="https://evemaps.dotlan.net/search?q={solar_system_name}" target="_blank" rel="noopener noreferer">{solar_system_name}</a>'.format(
-                solar_system_name=solar_system_name
+            solar_system_name_html = (
+                '<a href="https://evemaps.dotlan.net/search?q={solar_system_name}" '
+                'target="_blank" rel="noopener noreferer">{solar_system_name}</a>'.format(
+                    solar_system_name=solar_system_name
+                )
             )
 
             constellation_name = eve_solar_system.eve_constellation.name
-            constellation_name_html = '<a href="https://evemaps.dotlan.net/search?q={constellation_name}" target="_blank" rel="noopener noreferer">{constellation_name}</a>'.format(
-                constellation_name=constellation_name
+            constellation_name_html = (
+                '<a href="https://evemaps.dotlan.net/search?q={constellation_name}" '
+                'target="_blank" rel="noopener noreferer">{constellation_name}</a>'.format(
+                    constellation_name=constellation_name
+                )
             )
 
             region_name = eve_solar_system.eve_constellation.eve_region.name
-            region_name_html = '<a href="https://evemaps.dotlan.net/search?q={region_name}" target="_blank" rel="noopener noreferer">{region_name}</a>'.format(
-                region_name=region_name
+            region_name_html = (
+                '<a href="https://evemaps.dotlan.net/search?q={region_name}" '
+                'target="_blank" rel="noopener noreferer">{region_name}</a>'.format(
+                    region_name=region_name
+                )
             )
 
             structure_adm = 1
@@ -100,9 +113,32 @@ def dashboard_data(request) -> JsonResponse:
 
             start_time = campaign["start_time"].replace(tzinfo=None)
 
-            remaining = dt.timedelta(
+            remaining_time_in_seconds = dt.timedelta(
                 seconds=(start_time.timestamp() - dt.datetime.now().timestamp())
             ).total_seconds()
+
+            attackers_score = campaign["attackers_score"]
+            attackers_score_percent = "{:.0f}%".format(attackers_score * 100)
+            if (attackers_score * 100) < 10:
+                attackers_score_percent = "0" + attackers_score_percent
+
+            defender_score = campaign["defender_score"]
+            defender_score_percent = "{:.0f}%".format(defender_score * 100)
+            if (defender_score * 100) < 10:
+                defender_score_percent = "0" + defender_score_percent
+
+            active_campaign = _("No")
+            constellation_killboard_link = ""
+            if remaining_time_in_seconds < 0:
+                active_campaign = _("Yes")
+                constellation_killboard_link = (
+                    '<a href="https://zkillboard.com/constellation/{constellation_id}/" '
+                    'target="_blank" rel="noopener noreferer" '
+                    'class="aa-sov-timer-zkb-icon">{zkb_icon}</a>'.format(
+                        constellation_id=campaign["constellation_id"],
+                        zkb_icon='<img src="/static/sovtimer/images/zkillboard.png">',
+                    )
+                )
 
             data.append(
                 {
@@ -117,19 +153,17 @@ def dashboard_data(request) -> JsonResponse:
                     "structure_id": campaign["structure_id"],
                     "region_name": region_name,
                     "region_name_html": region_name_html,
-                    "attackers_score": "{:.0f}%".format(
-                        campaign["attackers_score"] * 100
-                    ),
+                    "attackers_score": attackers_score_percent,
                     "defender_id": campaign["defender_id"],
                     "defender_name": defender_name,
                     "defender_name_html": defender_name_html,
-                    "defender_score": "{:.0f}%".format(
-                        campaign["defender_score"] * 100
-                    ),
+                    "defender_score": defender_score_percent
+                    + constellation_killboard_link,
                     "adm": structure_adm,
                     "start_time": start_time,
                     "remaining_time": "",
-                    "remaining_time_in_seconds": remaining,
+                    "remaining_time_in_seconds": remaining_time_in_seconds,
+                    "active_campaign": active_campaign,
                 }
             )